création,modif de sauce ok et like dislike fait tout est bon il manque plus que les test

+++ backend/controllers/sauce.js
@@ -1,26 +1,64 @@
 const Sauce = require("../models/Sauces")
+const fs = require('fs');
+
+
+
 
 exports.createSauce = (req, res, next) => {
-    delete req.body._id
+    const sauceObject = JSON.parse(req.body.sauce);
+    delete sauceObject._id;
+    delete sauceObject._userId;
+    console.log(sauceObject)
     const sauce = new Sauce({
-        ...req.body
-
+        ...sauceObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
+
     sauce.save()
-        .then(() => res.status(201).json({ message: "sauce enregistrée" }))
-        .catch(error => res.status(400).json({ error }));
+        .then(() => { res.status(201).json({ message: 'sauce enregistré !' }) })
+        .catch(error => { res.status(400).json({ error }) })
 };
 
 exports.modifySauce = (req, res, next) => {
-    Sauce.updateOne({ _id: req.params.id }, { ...req.body.params.id })
-        .then(() => res.status(200).json({ message: "sauce modifié" }))
-        .catch(error => res.status(400).json({ error }))
+    console.log(req.body.sauce)
+    const sauceObject = req.file ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'sauce modifié!' }))
+                    .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
 };
 
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: "sauce supprimée" }))
-        .catch(error => res.status(400).json({ error }))
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: 'sauce supprimé !' }) })
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
 };
 
 exports.oneSauce = (req, res, next) => {
@@ -33,4 +71,91 @@ exports.allSauce = (req, res, next) => {
     Sauce.find()
         .then(sauces => res.status(200).json(sauces))
         .catch(error => res.status(400).json({ error }));
+}
+
+exports.noteSauce = (req, res, next) => {
+
+    //Si utilisateur like
+    
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            const saucetest = sauce.usersLiked.includes(req.body.userId)
+            if (req.body.like === 1) {
+                Sauce.updateOne({ _id: req.params.id },
+                    {
+                        $inc: { likes: 1 },
+                        $push: { usersLiked: req.body.userId },
+                    })
+                    .then(() => res.status(200).json({ message: "like!" }))
+                    .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+
+
+    //Si utilisateur annule son like
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            console.log(sauce.usersLiked)
+            const sauceLike = sauce.usersLiked.includes(req.body.userId)
+            console.log(sauceLike)
+            if (req.body.like === 0 && sauceLike === true) {
+                Sauce.updateOne({ _id: req.params.id },
+                    {
+                        $inc: { likes: -1 },
+                        $pull: { usersLiked: req.body.userId, }
+                    })
+                    .then(() => res.status(200).json({ message: "like enlevé!" }))
+                    .catch(error => res.status(401).json({ error }));
+
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+
+
+    // Si utilisateur dislike la sauce
+
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            if (req.body.like === -1) {
+                Sauce.updateOne({ _id: req.params.id },
+                    {
+                        $inc: { dislikes: 1 },
+                        $push: { usersDisliked: req.body.userId },
+                    })
+                    .then(() => res.status(200).json({ message: "dislikes!" }))
+                    .catch(error => res.status(401).json({ error }));
+
+
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+
+    //Si utilisateur annule son dislike
+
+    Sauce.findOne({ _id: req.params.id })
+        .then((sauce) => {
+            console.log(sauce.usersLiked)
+            const sauceDislike = sauce.usersDisliked.includes(req.body.userId)
+            console.log(sauceDislike)
+            if (req.body.like === 0 && sauceDislike === true) {
+                Sauce.updateOne({ _id: req.params.id },
+                    {
+                        $inc: { dislikes: -1 },
+                        $pull: { usersDisliked: req.body.userId, }
+                    })
+                    .then(() => res.status(200).json({ message: "dislikes enlevé!" }))
+                    .catch(error => res.status(401).json({ error }));
+
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
 }
